Change .inc/*inc.php file names to inc/*.php
This was redundant. By default, in inc/ directory each files are
designed to be included so it is not necessary to precise it with
a .inc.php extension. Therefore, *.class.php files remain named
with the .class extension to specified that they are Object Oriented.

Are inc/ files save against people trying to call them directly ?
e. g. if people request the page example.com/inc/foo.php.