Change login requiring behaviour.
Instead of drawing systematically the login page in case of authentification need, it is now the role of the theme to handle it.
This lead the theme to contain not only visual aspect but also a little bit of logic.

Note that in case of security need of user connection inside backend, you can use the `require_auth()` function from `function.php` to fail in case of no authentification.