Option to switch between display modes for entries working
Templates should use the `displayed_content` to get the user-preferred
way of displaying entries.

* If the user wants only the content, it contains the content, or the
description if content does not exist.
* If the user wants the description, it contains the description (or
first words of the content if there aren't any description).
* *Beware*, if the user wants to have only the titles, this variable
will be **empty**.

Templates still have access to `description`, `content` and `title` keys
in the entry array, to overwrite the global preferences.